requirement.txt
autopep8==1.7.0
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycodestyle==2.9.1
toml==0.10.2
Werkzeug==2.2.2
flask_sqlalchemy
flask_migrate